+++ cmd/mock.go
@@ -126,7 +126,7 @@ var ViperWriteConfigAsMock func(string) error
 
 var IsEqualMock func(arr1 []uint32, arr2 []uint32) (bool, int)
 
-var GetActiveAssetIdsMock func(*ethclient.Client, string) ([]uint8, error)
+var GetActiveAssetIdsMock func(*ethclient.Client, string) ([]uint16, error)
 
 var GetBlockManagerMock func(*ethclient.Client) *bindings.BlockManager
 
@@ -516,7 +516,7 @@ func (u UtilsMock) IsEqual(arr1 []uint32, arr2 []uint32) (bool, int) {
 	return IsEqualMock(arr1, arr2)
 }
 
-func (u UtilsMock) GetActiveAssetIds(client *ethclient.Client, address string) ([]uint8, error) {
+func (u UtilsMock) GetActiveAssetIds(client *ethclient.Client, address string) ([]uint16, error) {
 	return GetActiveAssetIdsMock(client, address)
 }
