+++ utils/array.go
@@ -70,10 +70,10 @@ func CalculateSumOfUint8Array(data []uint8) uint {
 	return sum
 }
 
-func ConvertUintArrayToUint8Array(uintArr []uint) []uint8 {
-	var arr []uint8
+func ConvertUintArrayToUint8Array(uintArr []uint) []uint16 {
+	var arr []uint16
 	for _, datum := range uintArr {
-		arr = append(arr, uint8(datum))
+		arr = append(arr, uint16(datum))
 	}
 	return arr
 }
